+++ IHocrXmlProcessor.cs
@@ -1,0 +1,7 @@
+using System.Xml.Linq;
+
+interface IHocrXmlProcessor
+{
+    void Init();
+    void Process(string hocrFile, XDocument hocrXml);
+}
